Switch to old.gov.ge
A new site is being launched, but is still is in testing mode. It looks
like the data is being loaded in by JSON via JS from URLs at
https://www.gov.ge/getGovMembers/en/1 and
https://www.gov.ge/getGovMembers/en/2, but I'm going ot hold off on
rewriting to use those until the site is more stable, in case that
changes.

For the meantime, the old site is available at old.gov.ge, and looks
like it's still being updated, rather than a static snapshot (as it has
stories dated today), so just switch the URLs to use that for now.